requirements.txt
certifi==2018.4.16
chardet==3.0.4
dj-config-url==0.1.1
Django==2.0.5
djangorestframework==3.8.2
flake8==3.5.0
gunicorn==19.8.1
idna==2.6
mccabe==0.6.1
psycopg2==2.7.4
pycodestyle==2.3.1
pyflakes==1.6.0
pytz==2018.4
requests==2.18.4
urllib3==1.22
